*** empty log message ***

+++ libppr/ppr_fnamef.c
@@ -64,7 +64,7 @@ void ppr_fnamef(char target[], const char pattern[], ...)
 			if(len > space_left)
 			    {
 			    *di = '\0';
-			    libppr_throw(EXCEPTION_BADUSAGE, function, "overflow: \"%s\" -> \"%s\" item %d: str=\"%s\", len=%ld, space_left=%ld", pattern, target, item_number, str, (long int)len, (long int)space_left);
+			    libppr_throw(EXCEPTION_BADUSAGE, function, "overflow: \"%s\" -> \"%s\" item %d: str=\"%s\", len=%d, space_left=%d", pattern, target, item_number, str, (int)len, (int)space_left);
 			    }
 
 			/* debug("y1 pattern=\"%s\", target=%p, di=%p, str=%p, str=\"%s\", space_left=%d", pattern, target, di, str, str, space_left); */
